refactor : Created a new Profile Service for user profile related business logic.
