+++ harmoniq_frontend/src/app/dashboard/dashboard.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { UserDto } from '../shared/Models/user/UserDto';
-import { AuthService } from '../shared/services/auth.service';
 import { Roles } from '../shared/Models/user/Roles';
+import { ProfileService } from '../shared/services/profile.service';
 
 @Component({
   selector: 'app-dashboard',
@@ -13,12 +13,12 @@ export class DashboardComponent implements OnInit {
   userData: UserDto | undefined;
   isGuest: boolean = false;
 
-  // injecting the necessary dependencies and fetching the user details
-  constructor(private authService: AuthService) {}
+  // injecting the necessary dependencies
+  constructor(private profileService: ProfileService) {}
 
   // Checking if the user is a guest user or not
   ngOnInit(): void {
-    this.userData = this.authService.getUser();
+    this.userData = this.profileService.getUser();
     this.isGuest = this.userData?.role === Roles.GUEST;
   }
 }
