+++ harmoniq_frontend/src/app/shared/guards/auth.guard.ts
@@ -1,15 +1,15 @@
 import { inject } from '@angular/core';
 import { CanActivateFn, Router } from '@angular/router';
-import { AuthService } from '../services/auth.service';
+import { ProfileService } from '../services/profile.service';
 
 // This guard checks if the user is logged in or not
 export const authGuard: CanActivateFn = (_route, _state) => {
   // Injecting the required dependencies
   const router = inject(Router);
-  const authService = inject(AuthService);
+  const profileService = inject(ProfileService);
 
   // Checking if the user is logged in or not
-  if (authService.getUser()) {
+  if (profileService.getUser()) {
     return true;
   } else {
     router.navigate(['/auth']);
