+++ harmoniq_frontend/src/app/shared/guards/guest.guard.ts
@@ -1,15 +1,15 @@
 import { inject } from '@angular/core';
 import { CanActivateFn, Router } from '@angular/router';
-import { AuthService } from '../services/auth.service';
 import { Roles } from '../Models/user/Roles';
+import { ProfileService } from '../services/profile.service';
 
 export const guestGuard: CanActivateFn = (_route, _state) => {
   // Injecting the required dependencies
   const router = inject(Router);
-  const authService = inject(AuthService);
+  const profileService = inject(ProfileService);
 
   // Checking if the user is a guest
-  if (authService.getUser()?.role === Roles.GUEST) {
+  if (profileService.getUser()?.role === Roles.GUEST) {
     router.navigate(['/dashboard/threads']);
     return false;
   } else {
