+++ harmoniq_frontend/src/app/shared/interfaces/ProfileInterface.ts
@@ -1,0 +1,14 @@
+import { Observable } from 'rxjs';
+import { AuthResponse } from '../Models/auth/AuthResponse';
+
+export interface ProfileInterface {
+  getUser(): AuthResponse | undefined;
+
+  getUserFromLocal(): AuthResponse | undefined;
+
+  getUserSubject(): Observable<AuthResponse | undefined>;
+
+  setUserInLocal(user: AuthResponse): void;
+
+  logout(): void;
+}
