+++ harmoniq_frontend/src/app/shared/services/auth.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 import { AuthInterface } from '../interfaces/AuthInterface';
-import { catchError, map, Observable, Subject, tap } from 'rxjs';
+import { catchError, map, Observable, tap } from 'rxjs';
 import { AuthResponse } from '../Models/auth/AuthResponse';
 import { HttpClient } from '@angular/common/http';
 import { ApiErrorHandlerService } from './api-error-handler.service';
@@ -10,48 +10,16 @@ import {
   LOGIN_ENDPOINT,
   REGISTER_ENDPOINT,
 } from '../constants/url-constants';
+import { ProfileService } from './profile.service';
 
 @Injectable({ providedIn: 'root' })
 export class AuthService implements AuthInterface {
-  // Api URL and local storage tokens
-  private USER_DATA_TOKEN = 'USER_DATA';
-
-  // User Subject which will transfer state data to all the required places
-  private user: AuthResponse | undefined = undefined;
-  private userSubject = new Subject<AuthResponse | undefined>();
-
   // Injecting the necessary dependencies
   constructor(
     private http: HttpClient,
-    private apiErrorHandler: ApiErrorHandlerService
-  ) {
-    this.user = this.getUserFromLocal();
-    this.userSubject.next(this.user);
-  }
-
-  // This function returns the current user data
-  getUser(): AuthResponse | undefined {
-    return this.user ? { ...this.user } : undefined;
-  }
-
-  // This function returns the current User Observable
-  getUserSubject(): Observable<AuthResponse | undefined> {
-    return this.userSubject.asObservable();
-  }
-
-  // This function returns the current stored user in the local storage
-  getUserFromLocal(): AuthResponse | undefined {
-    const data = localStorage.getItem(this.USER_DATA_TOKEN);
-    return data ? JSON.parse(data) : undefined;
-  }
-
-  // This function stores the user data to the local storage
-  setUserInLocal(user: AuthResponse): void {
-    this.user = user;
-    this.userSubject.next(this.getUser());
-
-    localStorage.setItem(this.USER_DATA_TOKEN, JSON.stringify(user));
-  }
+    private apiErrorHandler: ApiErrorHandlerService,
+    private profileService: ProfileService
+  ) {}
 
   // This function registers the user as a member
   registerMember(user: {
@@ -72,7 +40,7 @@ export class AuthService implements AuthInterface {
       .post<ResponseWrapper<AuthResponse>>(LOGIN_ENDPOINT, user)
       .pipe(
         map((response) => response.data),
-        tap((user: AuthResponse) => this.setUserInLocal(user)),
+        tap((user: AuthResponse) => this.profileService.setUserInLocal(user)),
         catchError(this.apiErrorHandler.handleApiError)
       );
   }
@@ -83,15 +51,8 @@ export class AuthService implements AuthInterface {
       .post<ResponseWrapper<AuthResponse>>(LOGIN_AS_GUEST, null)
       .pipe(
         map((response) => response.data),
-        tap((user: AuthResponse) => this.setUserInLocal(user)),
+        tap((user: AuthResponse) => this.profileService.setUserInLocal(user)),
         catchError(this.apiErrorHandler.handleApiError)
       );
-  }
-
-  // This function logs out the user
-  logout(): void {
-    this.user = undefined;
-    this.userSubject.next(undefined);
-    localStorage.removeItem(this.USER_DATA_TOKEN);
   }
 }
