+++ harmoniq_frontend/src/app/shared/services/comment.service.ts
@@ -1,12 +1,12 @@
 import { HttpClient, HttpHeaders } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { ApiErrorHandlerService } from './api-error-handler.service';
-import { AuthService } from './auth.service';
 import { CommentInterface } from '../interfaces/CommentInterface';
 import { catchError, map, Observable } from 'rxjs';
 import { CommentDto } from '../Models/comment/CommentDto';
 import { ResponseWrapper } from '../Models/common/ResponseWrapper';
 import { CREATE_COMMENT_ENDPOINT } from '../constants/url-constants';
+import { ProfileService } from './profile.service';
 
 @Injectable({ providedIn: 'root' })
 export class CommentService implements CommentInterface {
@@ -16,13 +16,13 @@ export class CommentService implements CommentInterface {
   constructor(
     private http: HttpClient,
     private apiErrorHandler: ApiErrorHandlerService,
-    authService: AuthService
+    profileService: ProfileService
   ) {
     // Storing the token in the variable
-    this.token = authService.getUser()?.token || 'Invalid Token';
+    this.token = profileService.getUser()?.token || 'Invalid Token';
 
     // Subscribing to the user changes
-    authService.getUserSubject().subscribe({
+    profileService.getUserSubject().subscribe({
       next: (user) => (this.token = user?.token || 'Invalid Token'),
     });
   }
